+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_it.dart
@@ -2143,4 +2143,11 @@ class DartPdfEditorLocalizationsIt extends DartPdfEditorLocalizations {
 
   @override
   String get splitFailed => 'Impossibile dividere questo PDF.';
+
+  @override
+  String get tbToolContentDelete => 'Cancella';
+
+  @override
+  String get tbTipContentDelete =>
+      'Elimina contenuto: trascina un rettangolo oppure fai clic sui vertici del poligono e doppio clic per terminare';
 }